Moved ControlScheme to Input namespace
